Add files game_logic and player
