+++ lib/player.rb
@@ -1,0 +1,9 @@
+# Players playing
+class Player
+  attr_accessor :name, :symbol
+
+  def initialize(name, symbol)
+    @name = name
+    @symbol = symbol
+  end
+end
